Move `buildCollectionTree` to collection utils
